+++ rocketgram/api/input_invoice_message_content.py
@@ -1,0 +1,39 @@
+# Copyright (C) 2015-2021 by Vd.
+# This file is part of Rocketgram, the modern Telegram bot framework.
+# Rocketgram is released under the MIT License (see LICENSE).
+
+
+from dataclasses import dataclass
+from typing import Optional, List
+
+from .input_message_content import InputMessageContent
+from .labeled_price import LabeledPrice
+
+
+@dataclass(frozen=True)
+class InputInvoiceMessageContent(InputMessageContent):
+    """\
+    Represents InputInvoiceMessageContent object:
+    https://core.telegram.org/bots/api#inputinvoicemessagecontent
+    """
+
+    title: str
+    description: str
+    payload: str
+    provider_token: str
+    currency: str
+    prices: List[LabeledPrice]
+    max_tip_amount: Optional[int] = None
+    suggested_tip_amounts: Optional[List[int]] = None
+    provider_data: Optional[str] = None
+    photo_url: Optional[str] = None
+    photo_size: Optional[int] = None
+    photo_width: Optional[int] = None
+    photo_height: Optional[int] = None
+    need_name: Optional[bool] = None
+    need_phone_number: Optional[bool] = None
+    need_email: Optional[bool] = None
+    need_shipping_address: Optional[bool] = None
+    send_phone_number_to_provider: Optional[bool] = None
+    send_email_to_provider: Optional[bool] = None
+    is_flexible: Optional[bool] = None